+++ __tests__/player.test.ts
@@ -4,17 +4,20 @@ import { PrismaClient } from "@prisma/client";
 import { initAgent } from "./helpers";
 import CONFIG from "@/config";
 import { Whatsapp } from "@/notification/whatsapp";
+import { PlayerServices } from "@/components/players/services";
+import { AuthServices } from "@/components/auth/services";
 
+const USER_AGENT = "jest_test";
 let agent: SuperAgentTest;
 let prisma: PrismaClient;
+let playerId: string;
+let agentAccessToken: string;
+let agentEmail: string;
 
-beforeAll(async () => {
-  prisma = new PrismaClient();
-  agent = await initAgent();
-});
+beforeAll(setUp);
+afterAll(cleanUp);
 
 describe("[UNIT] => PLAYERS ROUTER", () => {
-  let playerId: string;
   let playerAccessToken: string;
   const username = "jest_test" + Date.now();
   const email = username + "@test.com";
@@ -22,7 +25,12 @@ describe("[UNIT] => PLAYERS ROUTER", () => {
   const movile_number = "5490000000000";
 
   describe("POST: /players", () => {
-    it.only("Should create a player", async () => {
+    const mockCreateCasinoPlayer = jest.fn(async () => -420);
+    jest
+      .spyOn((PlayerServices as any).prototype, "createCasinoPlayer")
+      .mockImplementation(mockCreateCasinoPlayer);
+
+    it("Should create a player", async () => {
       const mockSend = jest.fn();
       jest.spyOn(Whatsapp, "send").mockImplementation(mockSend);
 
@@ -35,6 +43,7 @@ describe("[UNIT] => PLAYERS ROUTER", () => {
 
       const result = response.body.data;
 
+      expect(mockCreateCasinoPlayer).toHaveBeenCalledTimes(1);
       expect(mockSend).toHaveBeenCalledTimes(1);
       expect(response.status).toBe(CREATED);
       expect(Object.keys(result)).toEqual(["access", "refresh", "player"]);
@@ -113,15 +122,58 @@ describe("[UNIT] => PLAYERS ROUTER", () => {
       expect(response.body.data[0].msg).toBe("Usuario con ese email ya existe");
     });
 
-    it("Should return 400 ya_existe", async () => {
-      const response = await agent.post(`/app/${CONFIG.APP.VER}/players`).send({
-        username,
-        password,
-        email: "jest_test" + Date.now() + "@test.com",
-      });
+    // it("Should return 400 ya_existe", async () => {
+    //   const response = await agent.post(`/app/${CONFIG.APP.VER}/players`).send({
+    //     username,
+    //     password,
+    //     email: "jest_test" + Date.now() + "@test.com",
+    //   });
+
+    //   expect(response.status).toBe(BAD_REQUEST);
+    //   expect(response.body.code).toBe("ya_existe");
+    // });
+  });
+
+  describe("POST: /players/:id", () => {
+    it("Should update player details", async () => {
+      const response = await agent
+        .post(`/app/${CONFIG.APP.VER}/players/${playerId}`)
+        .set("Authorization", `Bearer ${agentAccessToken}`)
+        .set("User-Agent", USER_AGENT)
+        .send({
+          first_name: "Jest",
+          last_name: "Test",
+        });
+
+      expect(response.status).toBe(OK);
+      expect(response.body.data.first_name).toBe("Jest");
+      expect(response.body.data.last_name).toBe("Test");
+    });
+
+    it("Should return 400 unknown field", async () => {
+      const response = await agent
+        .post(`/app/${CONFIG.APP.VER}/players/${playerId}`)
+        .set("Authorization", `Bearer ${agentAccessToken}`)
+        .set("User-Agent", USER_AGENT)
+        .send({
+          unknown_field: "Jest",
+        });
 
       expect(response.status).toBe(BAD_REQUEST);
-      expect(response.body.code).toBe("ya_existe");
+      expect(response.body.data[0].type).toBe("unknown_fields");
+    });
+
+    it("Should return 400 email already in use", async () => {
+      const response = await agent
+        .post(`/app/${CONFIG.APP.VER}/players/${playerId}`)
+        .set("Authorization", `Bearer ${agentAccessToken}`)
+        .set("User-Agent", USER_AGENT)
+        .send({
+          email: agentEmail,
+        });
+
+      expect(response.status).toBe(BAD_REQUEST);
+      expect(response.body.data[0].msg).toBe("Ese email ya está en uso");
     });
   });
 
@@ -169,9 +221,60 @@ describe("[UNIT] => PLAYERS ROUTER", () => {
   });
 
   describe("GET: /players", () => {
-    it("Should return player info", async () => {
+    it("Should return list of players", async () => {
       const response = await agent
         .get(`/app/${CONFIG.APP.VER}/players`)
+        .set("Authorization", `Bearer ${agentAccessToken}`)
+        .set("User-Agent", USER_AGENT);
+
+      console.log("RESPONSE", response.body);
+      expect(response.status).toBe(OK);
+      expect(response.body.data).toBeInstanceOf(Array);
+      expect(Object.keys(response.body.data[0])).toEqual([
+        "id",
+        "panel_id",
+        "username",
+        "password",
+        "email",
+        "first_name",
+        "last_name",
+        "date_of_birth",
+        "movile_number",
+        "country",
+        "balance_currency",
+        "status",
+        "created_at",
+        "updated_at",
+      ]);
+    });
+
+    it.each`
+      field               | value    | message
+      ${"page"}           | ${"-1"}  | ${"page must be greater than 0"}
+      ${"items_per_page"} | ${"0"}   | ${"items_per_page must be greater than 1"}
+      ${"sort_column"}    | ${"foo"} | ${"Invalid sort_column"}
+      ${"sort_direction"} | ${"baz"} | ${"sort_direction must be 'asc' or 'desc'"}
+    `("Shloud return 400", async ({ field, value, message }) => {
+      const response = await agent
+        .get(`/app/${CONFIG.APP.VER}/players?${field}=${value}`)
+        .set("Authorization", `Bearer ${agentAccessToken}`)
+        .set("User-Agent", USER_AGENT);
+
+      expect(response.status).toBe(BAD_REQUEST);
+      expect(response.body.data[0].msg).toBe(message);
+    });
+
+    it("Should return 401", async () => {
+      const response = await agent.get(`/app/${CONFIG.APP.VER}/players`);
+
+      expect(response.status).toBe(UNAUTHORIZED);
+    });
+  });
+
+  describe("GET: /players/:id", () => {
+    it("Should return player info", async () => {
+      const response = await agent
+        .get(`/app/${CONFIG.APP.VER}/players/${playerId}`)
         .set("Authorization", `Bearer ${playerAccessToken}`);
 
       expect(response.status).toBe(OK);
@@ -179,7 +282,15 @@ describe("[UNIT] => PLAYERS ROUTER", () => {
     });
 
     it("Should return 401", async () => {
-      const response = await agent.get(`/app/${CONFIG.APP.VER}/players`);
+      const response = await agent.get(
+        `/app/${CONFIG.APP.VER}/players/${playerId}`,
+      );
+
+      expect(response.status).toBe(UNAUTHORIZED);
+    });
+
+    it("Should return 403", async () => {
+      const response = await agent.get(`/app/${CONFIG.APP.VER}/players/abcd`);
 
       expect(response.status).toBe(UNAUTHORIZED);
     });
@@ -193,4 +304,22 @@ async function checkPlayerRole(player_id: string) {
   });
   expect(player!.roles).toHaveLength(1);
   expect(player!.roles[0].name).toBe("player");
+}
+
+async function setUp() {
+  prisma = new PrismaClient();
+  agent = await initAgent();
+  const agentUser = await prisma.player.findFirst({
+    where: { roles: { some: { name: CONFIG.ROLES.AGENT } } },
+  });
+  if (!agentUser) throw new Error("No agent found");
+  agentEmail = agentUser.email;
+  const authServices = new AuthServices();
+  const { tokens } = await authServices.tokens(agentUser.id, USER_AGENT);
+  agentAccessToken = tokens.access;
+}
+
+async function cleanUp() {
+  await prisma.player.delete({ where: { id: playerId } });
+  await prisma.$disconnect();
 }