+++ src/types/request/players.ts
@@ -1,3 +1,5 @@
+import { Player } from "@prisma/client";
+
 export type getPlayerId = string;
 
 export type PlayerRequest = {
@@ -18,5 +20,20 @@ export type Credentials = {
 };
 
 export type PlayerUpdatableProps = {
-  password: string;
+  password?: string;
+  email?: string;
+  movile_number?: string;
+  first_name?: string;
+  last_name?: string;
+};
+
+export type PlayerUpdateRequest = {
+  email?: string;
+  movile_number?: string;
+  first_name?: string;
+  last_name?: string;
+};
+
+export type PlayerOrderBy = {
+  [key in keyof Player]?: "asc" | "desc";
 };